fixup! vmcheck: use --selinux-policy when committing

+++ tests/vmcheck/overlay.sh
@@ -47,13 +47,15 @@ if [ -d $INSTTREE/etc ]; then # on CentOS, the dbus service file is in /usr
 fi
 
 # ✀✀✀ BEGIN hack to get --selinux-policy (https://github.com/ostreedev/ostree/pull/1114) ✀✀✀
-# this is fine, rsync doesn't modify in place
-mount -o rw,remount /usr
-# don't overwrite /etc/ to not mess up 3-way merge
-rsync -rlv --exclude '/etc/' vmcheck/usr/ /usr/
+if ! ostree commit --help | grep -q -e --selinux-policy; then
+  # this is fine, rsync doesn't modify in place
+  mount -o rw,remount /usr
+  # don't overwrite /etc/ to not mess up 3-way merge
+  rsync -rlv --exclude '/etc/' vmcheck/usr/ /usr/
+fi
 # ✀✀✀ END hack to get --selinux-policy ✀✀✀
 
 ostree refs --delete vmcheck || true
-ostree commit -b vmcheck --selinux-policy vmcheck -s '' \
-  --tree=dir=vmcheck --link-checkout-speedup
+ostree commit -b vmcheck --link-checkout-speedup \
+  --selinux-policy=vmcheck --tree=dir=vmcheck
 ostree admin deploy vmcheck
